Fix bug; couldnt delete posts without a photo

+++ Project Hashtag/Pages/Post.cshtml.cs
@@ -179,22 +179,24 @@ namespace Project_Hashtag.Pages
                     return Forbid();
                 }
 
-                string relativePath = post.PictureUrl.Replace("/uploads/", string.Empty);
-                string imagePath = baseDirectoryPath + "\\" + relativePath;
+				if (post.PictureUrl != null)
+				{
+					string relativePath = post.PictureUrl.Replace("/uploads/", string.Empty);
+					string imagePath = baseDirectoryPath + "\\" + relativePath;
 
+					if (System.IO.File.Exists(imagePath))
+					{
+						// Delete the file
+						System.IO.File.Delete(imagePath);
+						Console.WriteLine("Image deleted successfully.");
+					}
+					else
+					{
+						Console.WriteLine("Image file not found.");
+					}
+				}
 
-                if (System.IO.File.Exists(imagePath))
-                {
-                    // Delete the file
-                    System.IO.File.Delete(imagePath);
-                    Console.WriteLine("Image deleted successfully.");
-                }
-                else
-                {
-                    Console.WriteLine("Image file not found.");
-                }
-
-                database.Posts.Remove(post);
+				database.Posts.Remove(post);
                 database.SaveChanges();
 
                 return RedirectToPage("Profile", new { userID = userID});
